refactor + auto engine building + remove cuda
