+++ requirements.txt
@@ -1,10 +1,4 @@
-einops>=0.8.0
-colored>=1.1.0
-polygraphy>=0.49.0
-tensorrt>=10.12.0
-cuda-python>=12.0.0
-requests>=2.31.0
-tqdm>=4.66.0
-onnx>=1.20.0
-onnxsim>=0.5.0
-torch>=2.9.0
+einops
+colored
+polygraphy
+tensorrt
