+++ trt_utilities.py
@@ -1,3 +1,20 @@
+#
+# Copyright 2022 The HuggingFace Inc. team.
+# SPDX-FileCopyrightText: Copyright (c) 1993-2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 import torch
 from torch.cuda import nvtx
 from collections import OrderedDict
@@ -16,7 +33,6 @@ import tensorrt as trt
 from logging import error, warning
 from tqdm import tqdm
 import copy
-import cuda.bindings.runtime as cudart
 
 TRT_LOGGER = trt.Logger(trt.Logger.ERROR)
 G_LOGGER.module_severity = G_LOGGER.ERROR
@@ -43,17 +59,6 @@ else:
 torch_to_numpy_dtype_dict = {
     value: key for (key, value) in numpy_to_torch_dtype_dict.items()
 }
-
-# https://github.com/Jeff-LiangF/streamv2v/blob/18c1a3bd56ff348d54a3300605936980bb13b03c/src/streamv2v/acceleration/tensorrt/utilities.py
-def CUASSERT(cuda_ret):
-    err = cuda_ret[0]
-    if err != cudart.cudaError_t.cudaSuccess:
-        raise RuntimeError(
-            f"CUDA ERROR: {err}, error code reference: https://nvidia.github.io/cuda-python/module/cudart.html#cuda.cudart.cudaError_t"
-        )
-    if len(cuda_ret) > 1:
-        return cuda_ret[1]
-    return None
 
 class TQDMProgressMonitor(trt.IProgressMonitor):
     def __init__(self):
@@ -125,6 +130,7 @@ class TQDMProgressMonitor(trt.IProgressMonitor):
             # There is no need to propagate this exception to TensorRT. We can simply cancel the build.
             return False
 
+
 class Engine:
     def __init__(
         self,
@@ -136,72 +142,24 @@ class Engine:
         self.buffers = OrderedDict()
         self.tensors = OrderedDict()
         self.cuda_graph_instance = None  # cuda graph
-        self.graph = None
 
     def __del__(self):
-        # Clean up CUDA graph resources
-        if hasattr(self, 'cuda_graph_instance') and self.cuda_graph_instance is not None:
-            try:
-                cudart.cudaGraphDestroy(self.cuda_graph_instance)
-            except Exception:
-                pass
-        if hasattr(self, 'graph') and self.graph is not None:
-            try:
-                cudart.cudaGraphDestroy(self.graph)
-            except Exception:
-                pass
-
-        if hasattr(self, 'engine'):
-            del self.engine
-        if hasattr(self, 'context'):
-            del self.context
-        if hasattr(self, 'tensors'):
-            for key in list(self.tensors.keys()):
-                del self.tensors[key]
-            del self.tensors
-        if hasattr(self, 'buffers'):
-            del self.buffers
-        if hasattr(self, 'inputs'):
-            del self.inputs
-        if hasattr(self, 'outputs'):
-            del self.outputs
+        del self.engine
+        del self.context
+        del self.buffers
+        del self.tensors
 
     def reset(self, engine_path=None):
-        # Clean up CUDA graph resources first
-        if hasattr(self, 'cuda_graph_instance') and self.cuda_graph_instance is not None:
-            try:
-                cudart.cudaGraphDestroy(self.cuda_graph_instance)
-            except Exception:
-                pass
-            self.cuda_graph_instance = None
-        if hasattr(self, 'graph') and self.graph is not None:
-            try:
-                cudart.cudaGraphDestroy(self.graph)
-            except Exception:
-                pass
-            self.graph = None
-
-        if hasattr(self, 'engine') and self.engine is not None:
-            del self.engine
-        if hasattr(self, 'context') and self.context is not None:
-            del self.context
-        if hasattr(self, 'tensors'):
-            for key in list(self.tensors.keys()):
-                del self.tensors[key]
-            del self.tensors
-        if hasattr(self, 'buffers'):
-            del self.buffers
-
-        self.engine = None
-        self.context = None
-        self.engine_path = engine_path if engine_path else self.engine_path
+        del self.engine
+        del self.context
+        del self.buffers
+        del self.tensors
+        self.engine_path = engine_path
 
         self.buffers = OrderedDict()
         self.tensors = OrderedDict()
         self.inputs = {}
         self.outputs = {}
-        self.cuda_graph_instance = None
-        self.graph = None
 
     def build(
         self,
@@ -214,7 +172,7 @@ class Engine:
         timing_cache=None,
         update_output_names=None,
     ):
-        print(f"Building TensorRT engine for {onnx_path}: {self.engine_path}")
+        # print(f"Building TensorRT engine for {onnx_path}: {self.engine_path}")
         p = [Profile()]
         if input_profile:
             p = [Profile() for i in range(len(input_profile))]
@@ -265,13 +223,10 @@ class Engine:
         return 0
 
     def load(self):
+        # print(f"Loading TensorRT engine: {self.engine_path}")
         self.engine = engine_from_bytes(bytes_from_path(self.engine_path))
 
     def activate(self, reuse_device_memory=None):
-        # If engine was reset, reload it
-        if self.engine is None:
-            self.load()
-
         if reuse_device_memory:
             self.context = self.engine.create_execution_context_without_device_memory()
         #    self.context.device_memory = reuse_device_memory
@@ -279,20 +234,6 @@ class Engine:
             self.context = self.engine.create_execution_context()
 
     def allocate_buffers(self, shape_dict=None, device="cuda"):
-        # Clean up CUDA graph resources since tensors will be recreated
-        if hasattr(self, 'cuda_graph_instance') and self.cuda_graph_instance is not None:
-            try:
-                cudart.cudaGraphDestroy(self.cuda_graph_instance)
-            except Exception:
-                pass
-            self.cuda_graph_instance = None
-        if hasattr(self, 'graph') and self.graph is not None:
-            try:
-                cudart.cudaGraphDestroy(self.graph)
-            except Exception:
-                pass
-            self.graph = None
-
         nvtx.range_push("allocate_buffers")
         for idx in range(self.engine.num_io_tensors):
             name = self.engine.get_tensor_name(idx)
@@ -312,32 +253,25 @@ class Engine:
         nvtx.range_pop()
 
     def infer(self, feed_dict, stream, use_cuda_graph=False):
+        nvtx.range_push("set_tensors")
         for name, buf in feed_dict.items():
             self.tensors[name].copy_(buf)
 
         for name, tensor in self.tensors.items():
             self.context.set_tensor_address(name, tensor.data_ptr())
-
-        if use_cuda_graph:
-            if self.cuda_graph_instance is not None:
-                CUASSERT(cudart.cudaGraphLaunch(self.cuda_graph_instance, stream.ptr))
-                CUASSERT(cudart.cudaStreamSynchronize(stream.ptr))
-            else:
-                # do inference before CUDA graph capture
-                noerror = self.context.execute_async_v3(stream.ptr)
-                if not noerror:
-                    raise ValueError("ERROR: inference failed.")
-                # capture cuda graph
-                CUASSERT(
-                    cudart.cudaStreamBeginCapture(stream.ptr, cudart.cudaStreamCaptureMode.cudaStreamCaptureModeGlobal)
-                )
-                self.context.execute_async_v3(stream.ptr)
-                self.graph = CUASSERT(cudart.cudaStreamEndCapture(stream.ptr))
-                self.cuda_graph_instance = CUASSERT(cudart.cudaGraphInstantiate(self.graph, 0))
-        else:
-            noerror = self.context.execute_async_v3(stream.ptr)
-            if not noerror:
-                raise ValueError("ERROR: inference failed.")
-
+        nvtx.range_pop()
+        nvtx.range_push("execute")
+        noerror = self.context.execute_async_v3(stream)
+        if not noerror:
+            raise ValueError("ERROR: inference failed.")
+        nvtx.range_pop()
         return self.tensors
 
+    def __str__(self):
+        out = ""
+        for opt_profile in range(self.engine.num_optimization_profiles):
+            for binding_idx in range(self.engine.num_bindings):
+                name = self.engine.get_binding_name(binding_idx)
+                shape = self.engine.get_profile_shape(opt_profile, name)
+                out += f"\t{name} = {shape}\n"
+        return out
